+++ xpasses/XR006/testdata/src/a/alias.go
@@ -7,9 +7,27 @@ import (
 )
 
 func falias() {
-	_ = s.Resource{ // want "resource should not include redundant Timeouts implementation"
+	_ = s.Resource{ // want "resource should not configure Timeouts.Create without Create implementation"
 		Timeouts: &s.ResourceTimeout{
 			Create: s.DefaultTimeout(time.Minute),
+		},
+	}
+
+	_ = s.Resource{ // want "resource should not configure Timeouts.Delete without Delete implementation"
+		Timeouts: &s.ResourceTimeout{
+			Delete: s.DefaultTimeout(time.Minute),
+		},
+	}
+
+	_ = s.Resource{ // want "resource should not configure Timeouts.Read without Read implementation"
+		Timeouts: &s.ResourceTimeout{
+			Read: s.DefaultTimeout(time.Minute),
+		},
+	}
+
+	_ = s.Resource{ // want "resource should not configure Timeouts.Update without Update implementation"
+		Timeouts: &s.ResourceTimeout{
+			Update: s.DefaultTimeout(time.Minute),
 		},
 	}
 }