+++ xpasses/XR006/testdata/src/a/comment_ignore_v2.go
@@ -13,4 +13,25 @@ func fcommentignore_v2() {
 			Create: schema.DefaultTimeout(time.Minute),
 		},
 	}
+
+	//lintignore:XR006
+	_ = schema.Resource{
+		Timeouts: &schema.ResourceTimeout{
+			Delete: schema.DefaultTimeout(time.Minute),
+		},
+	}
+
+	//lintignore:XR006
+	_ = schema.Resource{
+		Timeouts: &schema.ResourceTimeout{
+			Read: schema.DefaultTimeout(time.Minute),
+		},
+	}
+
+	//lintignore:XR006
+	_ = schema.Resource{
+		Timeouts: &schema.ResourceTimeout{
+			Update: schema.DefaultTimeout(time.Minute),
+		},
+	}
 }